requirements.txt
pytorch_lightning==1.6.0
rich
fire
transformers==4.5.1
sacrebleu==2.0.0
datasets==1.4.1